Better handling for `wickedd-nanny` and empty `members`
If the interface does not exist, try reloading `wickedd-nanny`.

If `--members` is empty, treat it as such instead of using the value of `0`
use the value of `None`. This fixes the conditional exit if `--members`
is not set for a bridge or bond.

Also add a deleter for `vlan_id`.