Fix wipes from in-memory
When running in-memory, such as a PXE booted hypervisor, the wipe script
is failing on its root disk saftey check. The script is attempting to
prevent wiping the root disk, but when running in-memory there is no
root disk. The script does not handle the situation when no root disk
resolves, and it gets confused causing an early exit before any wipe
commands are ran.