Adds back in netbooted disk install
This adds back in a new, modified version of the disk installation for
netbooted nodes.

Since the ISO is not available for a netbooted host, at least not
consistently across vendors, this leverages the in-RAM squash image.

USB/RemoteISO and netboot installs will use slightly different
procedures for now, the netboot install could be refitted to also work
for the USB/RemoteISO.. but since I've tested them both and am short on
tiem I'm leaving them as-is since they work.