MTL-2284 Fix missing `VMSTORE` mounts
The `mkfs.xfs` line was referencing an undefined variable `$target`,
this prevents the FSLabel from applying (as well as the filesystem).

Also, the letter index is off by 1. It starts at `b` instead of `a`.

This ensures the index doesn't use non-alpha characters, starts at
`a`, and (when referenced) the letters are capitalized.