The HMN was removed from the management VM
The HMN was removed from the management VM in a previous set of PRs for
Fawkes, however the `management-vm.sh` script was neglected.

Without this change, deploying the management-vm in Fawkes 0.0.1-alpha.5
(latest) fails when locating bond0.hmn0.

This removes `bond0.hmn0`, and bumps the `eth3` interface configuration
down to `eth2`.