Fix directory merge
Directories like `root` and `etc` are not merging at all when the
overlayFS is created.

For some reason, if the directories that need to merge are created on
the overlayFS without mounting an overlay and just mounting the
overlayFS (upperdir) the merge works.

If the overlay is mounted (lowerdir + upperdir + workdir), and things
are touched in the merged area, then the merge fails the next time it's
mounted. This makes no sense to me, but I just need this to work for
now.