Copy tar and/or extracted tar into bootable
This has been causing me grief every Fawkes install, I fail to capy the archive into my bootable.

The docs don't say to do this by hand because they aren't written for the scenario where you reinstall from a running hypervisor. The docs expect the tar to be available on a workstation or server in the system's peripheral.

This ensures the tar/extracted is included on the bootable, so I don't lose it when I wipe and reboot.